% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Outputs a list of 1-edit distance neighbors of a word.}
\usage{
get_neighbors(stimuli, database)
}
\arguments{
\item{stimuli}{A character vector containing words or nonwords.}

\item{database}{A dataframe that must contain a \code{Phono} column that contains phonological transcriptions.}
}
\value{
A dataframe with \code{stimuli} and list of its neighbors (if any).
}
\description{
Outputs a list of 1-edit distance neighbors of a word.
}
\examples{
# load a database first
# get_neighbors(stimuli = c('xbet', 'gEt', 'hWs', 'xgEnst'), database = data)

}
