% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfreq.R
\name{get_nfreq}
\alias{get_nfreq}
\title{Calculates the neighborhood frequency of a word based on Levenshtein edit distance of 1}
\usage{
get_nfreq(stimuli, database)
}
\arguments{
\item{stimuli}{A character vector containing words or nonwords to calculate the degree of.}

\item{database}{A dataframe that must contain a \code{Phono} column that contains phonological transcriptions and a \code{Frequency} column with log frequencies.}
}
\value{
A dataframe with \code{stimuli} and corresponding neighborhood frequency (average frequency of its neighbors).
}
\description{
Calculates the neighborhood frequency of a word based on Levenshtein edit distance of 1
}
\examples{
# load a database first
# get_nfreq(stimuli = c('xbet', 'gEt', 'hWs', 'xgEnst'), database = data)

}
